+++ cmd/client/cmd/getbalance.go
@@ -52,9 +52,9 @@ var getbalanceCmd = &cobra.Command{
 		}
 
 		if address == nil {
-			fmt.Printf("no account is provided. the coinbase balance: %s\n", amount)
+			fmt.Printf("no account is provided. the coinbase balance: %s\n", BigToDecimal(amount))
 		} else {
-			fmt.Printf("Account: %s\nBalance: %s\n", address.ToHex(), amount)
+			fmt.Printf("Account: %s\nBalance: %s\n", address.ToHex(), BigToDecimal(amount))
 		}
 	},
 }
