+++ cmd/client/cmd/simple_decimal.go
@@ -12,14 +12,21 @@ import (
 	"strings"
 )
 
-//BigToDecimalfl simply changes big int to float64 which will miss 0 in the last
-func BigToDecimalfl(amount *big.Int) string {
+//BigToDecimal simply changes big int to decimal which will miss additional 0 in the last
+func BigToDecimal(amount *big.Int) string {
 	base := pow.SeeleToCoin
 	var quotient = big.NewInt(0)
 	var mod = big.NewInt(0)
+	var numstr string
 	quotient.Div(amount, base)
 	mod.Mod(amount, base)
-	numstr := quotient.Text(10) + "." + fmt.Sprintf("%08s", mod.Text(10))
-	numstr = strings.TrimRight(numstr, "0")
+	modValue := mod.Text(10)
+	quotientValue := quotient.Text(10)
+	if strings.EqualFold(modValue, "0") {
+		numstr = quotientValue
+	} else {
+		numstr := quotientValue + "." + fmt.Sprintf("%08s", modValue)
+		numstr = strings.TrimRight(numstr, "0")
+	}
 	return numstr
 }
